pyproject.toml
[project]
name = "amplitude-data-wrapper"
version = "0.6.0"
description = "python wrapper for using the amplitude analytics and taxonomy APIs"
authors = [
    {name = "Tobias McVey", email = "tobias.mcvey@nav.no"},
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["amplitude"]
license = "MIT"
dependencies = [
    "aiohttp>=3.13.5",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.3",
]

[project.urls]
"Homepage" = "https://github.com/navikt/amplitude-data-wrapper"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
